Main added with the task , more description at the comment. Last version 25/11/17 18:20
The main function works with the task that enables if the player want to play vs human player OR vs AI player.
Need to check the main method of the AI player. I have already checked his helper function and it looks good. But I think that after the whole tests we will know better.